requirements.txt
moviepy~=1.0.3
tqdm~=4.65.0
numpy~=1.24.3
bs4~=0.0.1
beautifulsoup4~=4.12.2
torch
h5py~=3.9.0
sentence-transformers
pysrt
tensorboardX
pyaudioanalysis
pandas
matplotlib